chore: move text outside tab component folder
